requirements.txt
Flask==2.0.1
gunicorn==20.1.0
libsass==0.21.0
Flask-Assets==2.0
google-auth==2.0.0
Flask-SQLAlchemy==2.5.1